+++ pages/adMarketPlace/index.tsx
@@ -1,6 +1,11 @@
 import { useContext } from "react";
 import Header from "../../src/components/Header/header";
-import { Typography, List, Button } from "antd";
+
+// antd imports
+import { DownOutlined } from "@ant-design/icons";
+import { Button, Dropdown, Menu, Space, Radio, Typography, List } from "antd";
+import type { RadioChangeEvent } from "antd";
+
 import { useRouter } from "next/router";
 import {
   GetAllAsks,
@@ -144,7 +149,86 @@ const Adlist: React.FC<AdlistProp> = ({ onRentClick }) => {
     },
   });
 
+  // add useState hooks here
+  const [price, setPrice] = useState("100MATIC");
+  const [views, setViews] = useState("100kViews");
+
+  const onChangePrice = (e: RadioChangeEvent) => {
+    console.log("radio checked", e.target.value);
+    setPrice(e.target.value);
+  };
+
+  //  menu items for price dropdown filter
+  const priceFilterMenu = (
+    <Menu
+      items={[
+        {
+          label: (
+            <Radio.Group onChange={onChangePrice} value={price}>
+              <Space direction="vertical">
+                <Radio value={"100MATIC"}>100 MATIC and under</Radio>
+                <Radio value={"200MATIC"}>200 MATIC and under</Radio>
+                <Radio value={"300MATIC"}>500 MATIC and under</Radio>
+              </Space>
+            </Radio.Group>
+          ),
+          key: "1",
+        },
+      ]}
+    />
+  );
+
+  const onChangeViews = (e: RadioChangeEvent) => {
+    console.log("radio checked", e.target.value);
+    setViews(e.target.value);
+  };
+
+  // menu items for number of view dropdown filter
+  const viewsFilterMenu = (
+    <Menu
+      items={[
+        {
+          label: (
+            <Radio.Group onChange={onChangeViews} value={views}>
+              <Space direction="vertical">
+                <Radio value={"1000Views"}>1000 views and under</Radio>
+                <Radio value={"10kViews"}>10K views and under</Radio>
+                <Radio value={"50kViews"}>50K views and under</Radio>
+                <Radio value={"100kViews"}>100K views and under</Radio>
+                <Radio value={"1MViews"}>1 million+ views</Radio>
+              </Space>
+            </Radio.Group>
+          ),
+          key: "1",
+        },
+      ]}
+    />
+  );
+
   return (
+    <>
+     {/* start dropdowns */}
+     <div className="flex flex-row items-center mb-3">
+        <span>Filter by</span>
+        <Dropdown overlay={priceFilterMenu} trigger={["click"]}>
+          <a onClick={(e) => e.preventDefault()}>
+            <Space className="inline-block px-6 py-2 border shadow text-black font-medium text-xs leading-tight uppercase rounded-full ml-4">
+              Price
+              <DownOutlined />
+            </Space>
+          </a>
+        </Dropdown>
+        <Dropdown overlay={viewsFilterMenu} trigger={["click"]}>
+          <a onClick={(e) => e.preventDefault()}>
+            <Space className="inline-block px-6 py-2 border shadow text-black font-medium text-xs leading-tight uppercase rounded-full ml-4">
+              Views
+              <DownOutlined />
+            </Space>
+          </a>
+        </Dropdown>
+      </div>
+      {/* end dropdowns */}
+
     <List
       loading={isLoadingAllAsks}
       style={{
@@ -159,9 +243,9 @@ const Adlist: React.FC<AdlistProp> = ({ onRentClick }) => {
       renderItem={(item) => {
         return (
           <List.Item
-            extra={
-              <Button onClick={() => onRentClick(item)}>Rent Space</Button>
-            }
+          extra={
+            <Button onClick={() => onRentClick(item)}>Rent Space</Button>
+          }
           >
             <List.Item.Meta
               // avatar={<Avatar src="https://joeschmoe.io/api/v1/random" />}
@@ -172,6 +256,7 @@ const Adlist: React.FC<AdlistProp> = ({ onRentClick }) => {
         );
       }}
     />
+      </>
   );
 };
