+++ src/components/SongList/SongList.tsx
@@ -25,85 +25,11 @@ const SongList: React.FC<SongListProps> = ({ playSong }) => {
     },
   });
 
-  // add useState hooks here
-  const [price, setPrice] = useState("100MATIC");
-  const [views, setViews] = useState("100kViews");
-
-  const onChangePrice = (e: RadioChangeEvent) => {
-    console.log("radio checked", e.target.value);
-    setPrice(e.target.value);
-  };
-
-  //  menu items for price dropdown filter
-  const priceFilterMenu = (
-    <Menu
-      items={[
-        {
-          label: (
-            <Radio.Group onChange={onChangePrice} value={price}>
-              <Space direction="vertical">
-                <Radio value={"100MATIC"}>100 MATIC and under</Radio>
-                <Radio value={"200MATIC"}>200 MATIC and under</Radio>
-                <Radio value={"300MATIC"}>500 MATIC and under</Radio>
-              </Space>
-            </Radio.Group>
-          ),
-          key: "1",
-        },
-      ]}
-    />
-  );
-
-  const onChangeViews = (e: RadioChangeEvent) => {
-    console.log("radio checked", e.target.value);
-    setViews(e.target.value);
-  };
-
-  // menu items for number of view dropdown filter
-  const viewsFilterMenu = (
-    <Menu
-      items={[
-        {
-          label: (
-            <Radio.Group onChange={onChangeViews} value={views}>
-              <Space direction="vertical">
-                <Radio value={"1000Views"}>1000 views and under</Radio>
-                <Radio value={"10kViews"}>10K views and under</Radio>
-                <Radio value={"50kViews"}>50K views and under</Radio>
-                <Radio value={"100kViews"}>100K views and under</Radio>
-                <Radio value={"1MViews"}>1 million+ views</Radio>
-              </Space>
-            </Radio.Group>
-          ),
-          key: "1",
-        },
-      ]}
-    />
-  );
+  
 
   return (
     <>
-      {/* start dropdowns */}
-      <div className="flex flex-row items-center mb-3">
-        <span>Filter by</span>
-        <Dropdown overlay={priceFilterMenu} trigger={["click"]}>
-          <a onClick={(e) => e.preventDefault()}>
-            <Space className="inline-block px-6 py-2 border shadow text-black font-medium text-xs leading-tight uppercase rounded-full ml-4">
-              Price
-              <DownOutlined />
-            </Space>
-          </a>
-        </Dropdown>
-        <Dropdown overlay={viewsFilterMenu} trigger={["click"]}>
-          <a onClick={(e) => e.preventDefault()}>
-            <Space className="inline-block px-6 py-2 border shadow text-black font-medium text-xs leading-tight uppercase rounded-full ml-4">
-              Views
-              <DownOutlined />
-            </Space>
-          </a>
-        </Dropdown>
-      </div>
-      {/* end dropdowns */}
+     
 
       {/* songlist */}
       <List
